commented out move football

+++ keil/stepper.h
@@ -10,13 +10,13 @@ void initStepper(void);
 void setStepperSpeed(long whatSpeed);
 
 void step(int steps_to_move);
-void movefootball(int yards);
+void movefootball(int8_t yards);
 
 /**
  * Returns true if hits an endzone.
  * The ball will not move in this case.
  **/
-bool moveFootballForwardBy(int yards);
+bool moveFootballForwardBy(int8_t yards);
 void moveFootballToYardage(int yardage);
 int getYardage(void);
